List auto-renewable certificates; Migrate existing LE cert secrets to new certificate models (#2867)
Pre 1.4 era, LE certs are stored as secrets. In 1.4 we introduce new model to store certs. This PR adds a migration that collects all LE certs from secrets and transforms those into new cert models.

Added also flag on API & CLI to show whether or not a given cert is automatically renewable. Without that people might get the impression that all LE certs could be automatically renewed.

The included fixture cert is a valid LE cert but the private key is mangled so should not pose any issues. Anyone should not be able to use it without the working key.